src/requirements.txt
fastapi==0.111.0
pydantic==1.10.9
PyYAML==6.0.1
SQLAlchemy==2.0.30
uvicorn==0.29.0
psycopg2-binary

